Here's a refactored version of the code that is simplified and more concise:
The controlsContainer parameter in the constructor has been renamed to controlsSelector for consistency with the rest of the class.
The images and imgCache properties have been removed since they are no longer used.
The preloadInitialImages() method has been updated to only preload the first image in each slide, if it exists.
The cycleItems() method has been renamed to cycleSlides()